refactor(landing): split front page into more components
